Add GoogleAPI Service and UpdateFunction

+++ ModelsLib/DatabaseModels/InstKoordinatesDatabasemodel.cs
@@ -11,7 +11,7 @@ namespace ModelsLib.DatabaseModels
         public Guid Id { get; set; }
         public DateTime CreatedAt { get; set; }
         public DateTime LastChangedAt { get; set; }
-        public decimal? lat { get; set; }
-        public decimal? lng { get; set; }
+        public double? lat { get; set; }
+        public double? lng { get; set; }
     }
 }
